binance/ohlc_fetcher/requirements.txt
python-binance
pandas
python-dotenv
unittest